requirements.txt
apache-airflow-providers-amazon==7.3.0
apache-airflow-providers-postgres==5.6.0
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2 